refactor(portal-frontend): SPEC-PORTAL-TAXONOMY-SPLIT-001 commit 3 — extract _components/CoverageWidget.tsx
Pulls the 280-line CoverageWidget function (coverage bars + per-node
inline rename/delete + Suggest CTA gating) out of TaxonomyTab.tsx
into its own file. Same prop signature: coverage, activeNodeId,
onNodeClick, onSuggest, isSuggesting, isBackfilling, canEdit,
onRename, onDelete. Internal state (editingNodeId, editingName,
editingDescription, confirmDeleteId) stays inside the sub-component.

MAX_HEALTHY_NODE_COUNT moves with CoverageWidget — it's only used
inside that component.

New test file __tests__/CoverageWidget.test.tsx (12 tests, all green):
- edit-mode start / cancel / submit + singleton enforcement
  (clicking edit on node B closes node A's edit)
- delete-confirm start / cancel + onDelete invocation
- Suggest CTA gating: empty state (>=10 chunks), populated state
  (>=10 untagged AND >5% pct), MAX_HEALTHY_NODE_COUNT cap
- canEdit=false hides rename + delete affordances
- onNodeClick fires with the node id

Paraglide is not mocked; tests query against the source-language ('en')
strings ('Rename', 'Delete', 'Cancel', 'Save', 'Suggest categories',
etc.).

Files:
- new: _components/CoverageWidget.tsx (315 lines)
- new: _components/__tests__/CoverageWidget.test.tsx (334 lines)
- modified: _components/TaxonomyTab.tsx — 889 -> 604 lines (-285)
  - removed inline CoverageWidget function + MAX_HEALTHY_NODE_COUNT
  - added: import { CoverageWidget } from './CoverageWidget'
  - trimmed unused lucide-react imports (Pencil, Trash2) and the
    no-longer-needed TaxonomyCoverage type import

Gates:
- tsc -b --force: clean
- npx eslint on changed files: clean
- vitest run: 290/290 (was 278; +12 CoverageWidget tests)
